+++ src/main/java/at/hannibal2/skyhanni/features/rift/everywhere/motes/ShowMotesNpcSellPrice.kt
@@ -130,7 +130,7 @@ class ShowMotesNpcSellPrice {
         if (!RiftAPI.inRift()) return
         pattern.matchMatcher(event.message) {
             config.burgerStacks = group("amount").toInt()
-            chat("§6[SkyHanni] Set your McGrubber's burger stacks to ${group("amount")}.")
+            chat("Set your McGrubber's burger stacks to ${group("amount")}.")
         }
     }