+++ src/main/java/at/hannibal2/skyhanni/features/slayer/SlayerProfitTracker.kt
@@ -4,6 +4,7 @@ import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.config.Storage
 import at.hannibal2.skyhanni.data.SlayerAPI
 import at.hannibal2.skyhanni.events.GuiRenderEvent
+import at.hannibal2.skyhanni.events.LorenzChatEvent
 import at.hannibal2.skyhanni.events.PurseChangeCause
 import at.hannibal2.skyhanni.events.PurseChangeEvent
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
@@ -20,11 +21,14 @@ import at.hannibal2.skyhanni.utils.LorenzUtils.addAsSingletonList
 import at.hannibal2.skyhanni.utils.LorenzUtils.addSelector
 import at.hannibal2.skyhanni.utils.LorenzUtils.sortedDesc
 import at.hannibal2.skyhanni.utils.NEUInternalName
+import at.hannibal2.skyhanni.utils.NEUInternalName.Companion.asInternalName
 import at.hannibal2.skyhanni.utils.NEUItems.getNpcPriceOrNull
 import at.hannibal2.skyhanni.utils.NEUItems.getPriceOrNull
 import at.hannibal2.skyhanni.utils.NumberUtil
 import at.hannibal2.skyhanni.utils.NumberUtil.addSeparators
+import at.hannibal2.skyhanni.utils.SimpleTimeMark
 import at.hannibal2.skyhanni.utils.StringUtils
+import at.hannibal2.skyhanni.utils.StringUtils.matches
 import at.hannibal2.skyhanni.utils.StringUtils.removeColor
 import at.hannibal2.skyhanni.utils.jsonobjects.SlayerProfitTrackerItemsJson
 import at.hannibal2.skyhanni.utils.renderables.Renderable
@@ -32,10 +36,17 @@ import at.hannibal2.skyhanni.utils.tracker.SkyHanniTracker
 import at.hannibal2.skyhanni.utils.tracker.TrackerData
 import com.google.gson.annotations.Expose
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
+import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
 
 object SlayerProfitTracker {
     private val config get() = SkyHanniMod.feature.slayer.itemProfitTracker
+
+    private val diceRollChatPattern =
+        "§eYour §r§(5|6High Class )Archfiend Dice §r§erolled a §r§.(?<number>.)§r§e! Bonus: §r§.(?<hearts>.*)❤".toPattern()
+
+    private val ARCHFIEND_DICE = "ARCHFIEND_DICE".asInternalName()
+    private val HIGH_CLASS_ARCHFIEND_DICE = "HIGH_CLASS_ARCHFIEND_DICE".asInternalName()
 
     private var itemLogCategory = ""
     private var baseSlayerType = ""
@@ -52,7 +63,7 @@ object SlayerProfitTracker {
         }
 
         @Expose
-        var items: MutableMap<NEUInternalName, SlayerItemProfit> = HashMap()
+        var items: MutableMap<NEUInternalName, SlayerItem> = HashMap()
 
         @Expose
         var mobKillCoins: Long = 0
@@ -63,7 +74,7 @@ object SlayerProfitTracker {
         @Expose
         var slayerCompletedCount = 0
 
-        class SlayerItemProfit {
+        class SlayerItem {
             @Expose
             var internalName: NEUInternalName? = null
 
@@ -76,7 +87,7 @@ object SlayerProfitTracker {
             @Expose
             var hidden = false
 
-            override fun toString() = "SlayerItemProfit{" +
+            override fun toString() = "SlayerItem{" +
                 "internalName='" + internalName + '\'' +
                 ", timesDropped=" + timesDropped +
                 ", totalAmount=" + totalAmount +
@@ -84,7 +95,7 @@ object SlayerProfitTracker {
                 '}'
         }
 
-        override fun toString() = "SlayerProfitList{" +
+        override fun toString() = "SlayerProfitTracker.Data{" +
             "items=" + items +
             ", mobKillCoins=" + mobKillCoins +
             ", slayerSpawnCost=" + slayerSpawnCost +
@@ -135,10 +146,10 @@ object SlayerProfitTracker {
 
     private fun addItemPickup(internalName: NEUInternalName, stackSize: Int) {
         getTracker()?.modify {
-            val slayerItemProfit = it.items.getOrPut(internalName) { Data.SlayerItemProfit() }
+            val slayerItem = it.items.getOrPut(internalName) { Data.SlayerItem() }
 
-            slayerItemProfit.timesDropped++
-            slayerItemProfit.totalAmount += stackSize
+            slayerItem.timesDropped++
+            slayerItem.totalAmount += stackSize
         }
     }
 
@@ -183,7 +194,23 @@ object SlayerProfitTracker {
         if (!SlayerAPI.isInCorrectArea) return
         if (!SlayerAPI.hasActiveSlayerQuest()) return
 
-        addItem(event.internalName, event.amount)
+        val internalName = event.internalName
+        if (internalName == ARCHFIEND_DICE || internalName == HIGH_CLASS_ARCHFIEND_DICE) {
+            if (lastDiceRoll.passedSince() < 500.milliseconds) {
+                return
+            }
+        }
+
+        addItem(internalName, event.amount)
+    }
+
+    private var lastDiceRoll = SimpleTimeMark.farPast()
+
+    @SubscribeEvent
+    fun onChat(event: LorenzChatEvent) {
+        if (diceRollChatPattern.matches(event.message)) {
+            lastDiceRoll = SimpleTimeMark.now()
+        }
     }
 
     private fun addItem(internalName: NEUInternalName, amount: Int) {
@@ -196,7 +223,7 @@ object SlayerProfitTracker {
         addItemPickup(internalName, amount)
         logger.log("Coins gained for picking up an item ($itemName) ${price.addSeparators()}")
         if (config.priceInChat && price > config.minimumPrice) {
-            LorenzUtils.chat("§e[SkyHanni] §a+Slayer Drop§7: §r$itemName")
+            LorenzUtils.chat("§a+Slayer Drop§7: §r$itemName")
         }
         if (config.titleWarning && price > config.minimumPriceWarning) {
             LorenzUtils.sendTitle("§a+ $itemName", 5.seconds)
@@ -248,7 +275,7 @@ object SlayerProfitTracker {
 
                     if (KeyboardManager.isControlKeyDown()) {
                         itemLog.items.remove(internalName)
-                        LorenzUtils.chat("§e[SkyHanni] Removed $cleanName §efrom slayer profit display.")
+                        LorenzUtils.chat("Removed $cleanName §efrom slayer profit display.")
                         lastClickDelay = System.currentTimeMillis() + 500
                     } else {
                         itemProfit.hidden = !hidden
@@ -337,9 +364,9 @@ object SlayerProfitTracker {
 
     fun clearProfitCommand(args: Array<String>) {
         if (itemLogCategory == "") {
-            LorenzUtils.chat(
-                "§c[SkyHanni] No current slayer data found. " +
-                    "Go to a slayer area and start the specific slayer type you want to reset the data of."
+            LorenzUtils.userError(
+                "No current slayer data found! " +
+                    "§eGo to a slayer area and start the specific slayer type you want to reset the data of.",
             )
             return
         }